+++ src/main/java/wooteco/subway/controller/dto/request/StationRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station;
+package wooteco.subway.controller.dto.request;
 
 public class StationRequest {
     private String name;
