+++ src/main/java/wooteco/subway/domain/Line.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line;
+package wooteco.subway.domain;
 
 public class Line {
     private final Long id;
