+++ src/main/java/wooteco/subway/service/LineService.java
@@ -1,6 +1,11 @@
-package wooteco.subway.line;
+package wooteco.subway.service;
 
 import org.springframework.stereotype.Service;
+import wooteco.subway.controller.dto.request.LineEditRequest;
+import wooteco.subway.controller.dto.request.LineRequest;
+import wooteco.subway.controller.dto.response.LineResponse;
+import wooteco.subway.dao.LineDao;
+import wooteco.subway.domain.Line;
 
 import java.util.List;
 import java.util.stream.Collectors;
